Refactoring and update README.md
