+++ .idea/csv-editor.xml
@@ -10,13 +10,6 @@
             </Attribute>
           </value>
         </entry>
-        <entry key="/pyPromptChecker-bin/model_list.csv">
-          <value>
-            <Attribute>
-              <option name="separator" value="," />
-            </Attribute>
-          </value>
-        </entry>
       </map>
     </option>
   </component>
