+++ config.ini
@@ -41,7 +41,7 @@ JsonMultiple=directory
 # If 'JsonSingle=filename' is set, the image file name will be set as default.
 # If 'JsonMultiple=directory' is set, the first image's directory name will be set as default.
 
-MoveDelete=False
+MoveDelete=True
 UseCopyInsteadOfMove=True
 AskIfDelete=True
 AskIfClearTrashBin=True
@@ -67,7 +67,7 @@ ModelHashExtractor=False
 # ModelHashExtractor is toggle shown/not shown 'M' button.
 # If 'False' here 'M' button is not shown.
 
-TabNavigation=False
+TabNavigation=True
 
 # If set to True here and  more than 5 files are loaded,
 # 'jump to' button and tab selection box appear at the top of the window.
