+++ pyPromptChecker/main.py
@@ -4,10 +4,9 @@ import os
 import sys
 import argparse
 import glob
-import pyPromptChecker
 
-from pyPromptChecker.lib import decoder
-from pyPromptChecker.lib import window
+from lib import decoder
+from gui import window
 
 
 def directory_to_filelist(directory_path):
@@ -30,10 +29,10 @@ def check_files(target_list):
     valid_file_list = []
 
     if file_counts > 20:
-        app, progress_bar = pyPromptChecker.lib.window.from_main(False, True)
+        app, progress_bar = window.from_main('progress')
         progress_bar.setLabelText("Checking files...")
         progress_bar.setRange(0, file_counts)
-        pyPromptChecker.lib.window.move_center(progress_bar)
+        window.move_center(progress_bar)
         progress_enable = True
 
     for filepath in target_list:
@@ -49,7 +48,7 @@ def check_files(target_list):
             progress_bar.update_value()
 
     if progress_enable:
-        pyPromptChecker.lib.window.from_main(False, True, progress_bar)
+        progress_bar.close()
 
     return valid_file_list, file_is_not_found_list, this_is_directory_list, this_file_is_not_png_file_list
 
@@ -70,10 +69,10 @@ def main():
     elif args.directory:
         filepaths = directory_to_filelist(args.directory)
     elif args.ask:
-        src = pyPromptChecker.lib.window.from_main(True)
+        src = window.from_main('directory')
         filepaths = directory_to_filelist(src) if src else None
     else:
-        filepaths = pyPromptChecker.lib.window.from_main()
+        filepaths = window.from_main('files')
 
     if filepaths:
         valid_filepath, not_found_list, directory_list, not_png_list = check_files(filepaths)
@@ -96,7 +95,7 @@ def main():
             sys.exit()
         print('a hoy!!!!')
         valid_filepath.sort()
-        window.result_window(valid_filepath)
+        window.from_main('result', valid_filepath)
 
 
 if __name__ == '__main__':
